rename modules in misc.latex
