+++ pyphs/misc/latex/__init__.py
@@ -1,7 +1,7 @@
 #
 
-from .core2tex import core2tex
-from .netlist2tex import netlist2tex, graphplot2tex
+from .core import core2tex
+from .netlist import netlist2tex, graphplot2tex
 from .document import texdocument
 
 __all__ = ['core2tex', 'netlist2tex', 'graphplot2tex', 'texdocument']
